+++ src/home/HomeMenuBar.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Nav, Navbar, NavItem } from 'react-bootstrap';
-import styles from '../../../public/css/index.css';
+import styles from '../../public/css/index.css';
 
 export class HomeMenuBar extends React.Component {
   constructor(props) {
